+++ components/answer-sheet-builder/components/form/LineStylePicker.tsx
@@ -4,7 +4,10 @@ import { useCallback, useEffect, useRef, useState } from "react"
 
 import { Slider } from "@/components/ui/slider"
 import { cn } from "@/lib/utils"
-import type { BorderConfig, LineStyle } from "@/types/answerSheetBuilder.types"
+import type {
+  BorderConfig,
+  LineStyle,
+} from "@/types/answerSheetDefinition.types"
 
 interface LineStylePickerProps {
   borderConfig: BorderConfig
