+++ components/answer-sheet-builder/hooks/useAnswerSheetExport.ts
@@ -7,14 +7,14 @@
 import { useCallback, useState } from "react"
 import { toast } from "sonner"
 
-import type { AnswerSheetDefinition } from "@/types/answerSheetBuilder.types"
+import type { AnswerSheetDefinition } from "@/types/answerSheetDefinition.types"
 
 import {
   renderMultiPageSvgStrings,
   resolveImageDataUris,
   wrapSvgsInHtml,
 } from "../utils/renderSvgStrings"
-import { computeMultiPageLayoutFromDefinition } from "./useAnswerSheetLayout"
+import { computeMultiPageLayoutFromDefinition } from "./layout/computeMultiPageLayout"
 
 export function useAnswerSheetExport() {
   const [isExporting, setIsExporting] = useState(false)
